Abstract Main into BobbyBot to handle user inputs
